requirements.txt
watchdog
google-cloud-vision
google-cloud
